+++ cli/src/commands/workspace/mod.rs
@@ -15,6 +15,7 @@
 mod add;
 mod forget;
 mod list;
+mod remove;
 mod rename;
 mod root;
 mod update_stale;
@@ -29,6 +30,8 @@ use self::forget::WorkspaceForgetArgs;
 use self::forget::cmd_workspace_forget;
 use self::list::WorkspaceListArgs;
 use self::list::cmd_workspace_list;
+use self::remove::WorkspaceRemoveArgs;
+use self::remove::cmd_workspace_remove;
 use self::rename::WorkspaceRenameArgs;
 use self::rename::cmd_workspace_rename;
 use self::root::WorkspaceRootArgs;
@@ -55,6 +58,7 @@ pub(crate) enum WorkspaceCommand {
     Add(WorkspaceAddArgs),
     Forget(WorkspaceForgetArgs),
     List(WorkspaceListArgs),
+    Remove(WorkspaceRemoveArgs),
     Rename(WorkspaceRenameArgs),
     Root(WorkspaceRootArgs),
     UpdateStale(WorkspaceUpdateStaleArgs),
@@ -70,6 +74,7 @@ pub(crate) async fn cmd_workspace(
         WorkspaceCommand::Add(args) => cmd_workspace_add(ui, command, args).await,
         WorkspaceCommand::Forget(args) => cmd_workspace_forget(ui, command, args).await,
         WorkspaceCommand::List(args) => cmd_workspace_list(ui, command, args).await,
+        WorkspaceCommand::Remove(args) => cmd_workspace_remove(ui, command, args).await,
         WorkspaceCommand::Rename(args) => cmd_workspace_rename(ui, command, args).await,
         WorkspaceCommand::Root(args) => cmd_workspace_root(ui, command, args).await,
         WorkspaceCommand::UpdateStale(args) => cmd_workspace_update_stale(ui, command, args).await,